+++ statement.c
@@ -310,7 +310,7 @@ int eval_statement(struct statement *s, int number, int next_number) {
 		case REM:
 			break;
 		case RANDOMIZE:
-			tcb_srand(time(NULL));
+			tcb_srand((unsigned long) time(NULL));
 			break;
 	}
